Fix test imports after error folder restructure
