+++ test/tests/unit/set-user-unit.spec.ts
@@ -2,8 +2,8 @@ import { describe, it, expect, beforeEach } from 'vitest'
 import { SetUserUnitService, UnitNotFromOrganizationError } from '../../../src/services/users/set-user-unit'
 import { InMemoryUserRepository } from '../../../src/repositories/in-memory/in-memory-users-repository'
 import { FakeUnitRepository } from '../../helpers/fake-repositories'
-import { UserNotFoundError } from '../../../src/services/@errors/user-not-found-error'
-import { UnitNotFoundError } from '../../../src/services/@errors/unit-not-found-error'
+import { UserNotFoundError } from '../../../src/services/@errors/user/user-not-found-error'
+import { UnitNotFoundError } from '../../../src/services/@errors/unit/unit-not-found-error'
 import { namedUser, makeUnit } from '../../helpers/default-values'
 
 const user = { ...namedUser }
